// src/lib/sanity/sanity.types.ts
export type SiteSettingsQueryResult =
  | {
      title: string | null;
      description: null;
      entityType: null;
      entityName: null;
      footerText: null;
      navigation: null;
      social: null;
    }
  | {
      title: string | null;
      description: string | null;
      entityType: null;
      entityName: null;
      footerText: null;
      navigation: null;
      social: null;
    }
  | {
      title: string | null;
      description: string | null;
      entityType: "organization" | "person" | null;
      entityName: string | null;
      footerText: string | null;
      navigation: Array<{
        label: string | null;
        href: string | null;
      }> | null;
      social: Array<{
        label: string | null;
        url: string | null;
      }> | null;
    }
  | null;